+++ docs/guide/source/changelog.rst
@@ -17,6 +17,14 @@ This also includes breaking changes that MuJoCo itself introduced, thus even an
 update of MuJoCo alone can increase the major version.
 
 
+Unreleased (MuJoCo 3.3.7)
+=============================
+- Changes to :ref:`mj_rust_viewer`:
+
+  - Added :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>add_ui_callback_detached` for custom UI widgets that
+    do not require the simulation state (i.e., the passive |mj_data| instance).
+
+
 2.2.0 (MuJoCo 3.3.7)
 ================================
 - Made :ref:`mj_renderer` use EGL backend by default on Linux for true offscreen rendering (pbuffer).